requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.1
botocore==1.38.1
celery==5.5.2
cffi==2.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==46.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
kombu==5.5.3
llvmlite==0.44.0
numba==0.61.2
numpy==2.2.4
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.23
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
